// components/layout/Flex/Flex.d.ts
import { default as React } from '../../../../node_modules/react';
import { WimSpacing } from '../../../types/tokens';
export type FlexProps = React.ComponentPropsWithoutRef<"div"> & {
    /**
     * If true, the flex will be rendered as its child, merging its props onto that child.
     */
    asChild?: boolean;
    /** Flex direction */
    direction?: "row" | "row-reverse" | "column" | "column-reverse";
    /** Align-items */
    align?: "start" | "center" | "end" | "stretch" | "baseline";
    /** Justify-content */
    justify?: "start" | "center" | "end" | "between" | "around" | "evenly" | "stretch";
    /** Flex-wrap */
    wrap?: "nowrap" | "wrap" | "wrap-reverse";
    /** Gap between children */
    gap?: number | WimSpacing;
    /** Display inline-flex */
    inline?: boolean;
};
/**
 * Flex component for flexible layouts.
 */
export declare const Flex: React.ForwardRefExoticComponent<Omit<React.DetailedHTMLProps<React.HTMLAttributes<HTMLDivElement>, HTMLDivElement>, "ref"> & {
    /**
     * If true, the flex will be rendered as its child, merging its props onto that child.
     */
    asChild?: boolean;
    /** Flex direction */
    direction?: "row" | "row-reverse" | "column" | "column-reverse";
    /** Align-items */
    align?: "start" | "center" | "end" | "stretch" | "baseline";
    /** Justify-content */
    justify?: "start" | "center" | "end" | "between" | "around" | "evenly" | "stretch";
    /** Flex-wrap */
    wrap?: "nowrap" | "wrap" | "wrap-reverse";
    /** Gap between children */
    gap?: number | WimSpacing;
    /** Display inline-flex */
    inline?: boolean;
} & React.RefAttributes<HTMLDivElement>>;